Move links out of config
