+++ airbyte-webapp/src/packages/cloud/views/auth/SignupPage/components/SignupForm.tsx
@@ -8,11 +8,11 @@ import * as yup from "yup";
 import { LabeledInput, Link } from "components";
 import { Button } from "components/ui/Button";
 
-import { useConfig } from "config";
 import { useExperiment } from "hooks/services/Experiment";
 import { FieldError } from "packages/cloud/lib/errors/FieldError";
 import { useAuthService } from "packages/cloud/services/auth/AuthService";
 import { isGdprCountry } from "utils/dataPrivacy";
+import { links } from "utils/links";
 
 import CheckBoxControl from "../../components/CheckBoxControl";
 import { BottomBlock, FieldItem, Form, RowFieldItem } from "../../components/FormComponents";
@@ -134,19 +134,18 @@ export const NewsField: React.FC = () => {
 };
 
 export const Disclaimer: React.FC = () => {
-  const config = useConfig();
   return (
     <div className={styles.disclaimer}>
       <FormattedMessage
         id="login.disclaimer"
         values={{
           terms: (terms: React.ReactNode) => (
-            <Link $clear target="_blank" href={config.links.termsLink} as="a">
+            <Link $clear target="_blank" href={links.termsLink} as="a">
               {terms}
             </Link>
           ),
           privacy: (privacy: React.ReactNode) => (
-            <Link $clear target="_blank" href={config.links.privacyLink} as="a">
+            <Link $clear target="_blank" href={links.privacyLink} as="a">
               {privacy}
             </Link>
           ),
