+++ airbyte-webapp/src/packages/cloud/views/auth/components/GitBlock/GitBlock.tsx
@@ -3,7 +3,8 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React, { FC } from "react";
 import { FormattedMessage } from "react-intl";
 
-import { useConfig } from "../../../../../../config";
+import { links } from "utils/links";
+
 import styles from "./GitBlock.module.scss";
 
 export interface GitBlockProps {
@@ -11,10 +12,9 @@ export interface GitBlockProps {
   messageStyle?: React.CSSProperties;
 }
 export const GitBlock: FC<GitBlockProps> = ({ titleStyle, messageStyle }) => {
-  const config = useConfig();
   return (
     <div className={styles.container}>
-      <a className={styles.link} href={config.links.gitLink} target="_blank" rel="noreferrer">
+      <a className={styles.link} href={links.gitLink} target="_blank" rel="noreferrer">
         <div className={styles.content}>
           <FontAwesomeIcon icon={faGithub} className={styles.icon} />
           <div>
