+++ airbyte-webapp/src/pages/OnboardingPage/components/WelcomeStep.tsx
@@ -4,7 +4,7 @@ import styled from "styled-components";
 
 import { Button } from "components/ui/Button";
 
-import { useConfig } from "config";
+import { links } from "utils/links";
 
 import HighlightedText from "./HighlightedText";
 import TitlesBlock from "./TitlesBlock";
@@ -26,8 +26,6 @@ const Videos = styled.div`
 `;
 
 const WelcomeStep: React.FC<WelcomeStepProps> = ({ userName, onNextStep }) => {
-  const config = useConfig();
-
   return (
     <>
       <TitlesBlock
@@ -65,7 +63,7 @@ const WelcomeStep: React.FC<WelcomeStepProps> = ({ userName, onNextStep }) => {
         <VideoItem
           description={<FormattedMessage id="onboarding.exploreDemo" />}
           img="/videoCover.png"
-          link={config.links.demoLink}
+          link={links.demoLink}
         />
       </Videos>
       <Button size="lg" onClick={onNextStep}>
