+++ airbyte-webapp/src/pages/SettingsPage/pages/ConnectorsPage/components/CreateConnectorModal.tsx
@@ -8,7 +8,7 @@ import { LabeledInput, Link, StatusIcon } from "components";
 import { Button } from "components/ui/Button";
 import { Modal } from "components/ui/Modal";
 
-import { useConfig } from "config";
+import { links } from "utils/links";
 
 import styles from "./CreateConnectorModal.module.scss";
 
@@ -90,7 +90,6 @@ const validationSchema = yup.object().shape({
 });
 
 const CreateConnectorModal: React.FC<IProps> = ({ onClose, onSubmit, errorMessage }) => {
-  const config = useConfig();
   const { formatMessage } = useIntl();
 
   return (
@@ -101,7 +100,7 @@ const CreateConnectorModal: React.FC<IProps> = ({ onClose, onSubmit, errorMessag
             id="admin.learnMore"
             values={{
               lnk: (lnk: React.ReactNode) => (
-                <DocLink target="_blank" href={config.links.docsLink} as="a">
+                <DocLink target="_blank" href={links.docsLink} as="a">
                   {lnk}
                 </DocLink>
               ),
