+++ airbyte-webapp/src/views/Connection/ConnectionForm/components/SyncCatalogField.tsx
@@ -10,11 +10,11 @@ import { CheckBox } from "components/ui/CheckBox";
 import { LoadingBackdrop } from "components/ui/LoadingBackdrop";
 import { InfoTooltip, TooltipLearnMoreLink } from "components/ui/Tooltip";
 
-import { useConfig } from "config";
 import { SyncSchemaStream } from "core/domain/catalog";
 import { DestinationSyncMode } from "core/request/AirbyteClient";
 import { BatchEditProvider, useBulkEdit } from "hooks/services/BulkEdit/BulkEditService";
 import { ConnectionFormMode, useConnectionFormService } from "hooks/services/ConnectionForm/ConnectionFormService";
+import { links } from "utils/links";
 import { naturalComparatorBy } from "utils/objects";
 import CatalogTree from "views/Connection/CatalogTree";
 
@@ -66,7 +66,6 @@ interface SchemaViewProps extends FieldProps<SyncSchemaStream[]> {
 
 const CatalogHeader: React.FC = () => {
   const { mode } = useConnectionFormService();
-  const config = useConfig();
   const { onCheckAll, selectedBatchNodeIds, allChecked } = useBulkEdit();
   const catalogHeaderStyle = classnames({
     [styles.catalogHeader]: mode !== "readonly",
@@ -99,7 +98,7 @@ const CatalogHeader: React.FC = () => {
         <FormattedMessage id="form.syncMode" />
         <InfoTooltip>
           <FormattedMessage id="connectionForm.syncType.info" />
-          <TooltipLearnMoreLink url={config.links.syncModeLink} />
+          <TooltipLearnMoreLink url={links.syncModeLink} />
         </InfoTooltip>
       </Cell>
       <Cell lighter>
