+++ airbyte-webapp/src/views/Settings/PreferencesForm/PreferencesForm.tsx
@@ -9,7 +9,7 @@ import LabeledInput from "components/LabeledInput";
 import { LabeledSwitch } from "components/LabeledSwitch";
 import { Button } from "components/ui/Button";
 
-import { useConfig } from "config";
+import { links } from "utils/links";
 
 import EditControls from "./components/EditControls";
 
@@ -70,7 +70,6 @@ const PreferencesForm: React.FC<PreferencesFormProps> = ({
   errorMessage,
 }) => {
   const { formatMessage } = useIntl();
-  const config = useConfig();
 
   return (
     <Formik
@@ -125,7 +124,7 @@ const PreferencesForm: React.FC<PreferencesFormProps> = ({
               id="preferences.collectData"
               values={{
                 docs: (docs: React.ReactNode) => (
-                  <DocsLink target="_blank" href={config.links.docsLink}>
+                  <DocsLink target="_blank" href={links.docsLink}>
                     {docs}
                   </DocsLink>
                 ),
