+++ airbyte-webapp/src/views/layout/SideBar/SideBar.tsx
@@ -12,6 +12,7 @@ import Version from "components/Version";
 
 import { useConfig } from "config";
 import { useCurrentWorkspace } from "hooks/services/useWorkspace";
+import { links } from "utils/links";
 
 import { RoutePaths } from "../../../pages/routePaths";
 import ConnectionsIcon from "./components/ConnectionsIcon";
@@ -86,7 +87,7 @@ const SideBar: React.FC = () => {
       </div>
       <ul className={styles.menu}>
         <li>
-          <a href={config.links.updateLink} target="_blank" rel="noreferrer" className={styles.menuItem}>
+          <a href={links.updateLink} target="_blank" rel="noreferrer" className={styles.menuItem}>
             <FontAwesomeIcon className={styles.helpIcon} icon={faRocket} />
             <Text className={styles.text} size="sm">
               <FormattedMessage id="sidebar.update" />
@@ -102,19 +103,19 @@ const SideBar: React.FC = () => {
             options={[
               {
                 type: SidebarDropdownMenuItemType.LINK,
-                href: config.links.docsLink,
+                href: links.docsLink,
                 icon: <DocsIcon />,
                 displayName: <FormattedMessage id="sidebar.documentation" />,
               },
               {
                 type: SidebarDropdownMenuItemType.LINK,
-                href: config.links.slackLink,
+                href: links.slackLink,
                 icon: <FontAwesomeIcon icon={faSlack} />,
                 displayName: <FormattedMessage id="sidebar.joinSlack" />,
               },
               {
                 type: SidebarDropdownMenuItemType.LINK,
-                href: config.links.recipesLink,
+                href: links.recipesLink,
                 icon: <RecipesIcon />,
                 displayName: <FormattedMessage id="sidebar.recipes" />,
               },
